+++ api/src/main/java/io/devarium/api/auth/filter/JwtAuthenticationFilter.java
@@ -1,8 +1,8 @@
 package io.devarium.api.auth.filter;
 
-import io.devarium.infrastructure.security.jwt.constants.JwtConstants;
-import io.devarium.infrastructure.security.jwt.util.JwtUtil;
-import io.devarium.infrastructure.security.userdetails.CustomUserDetailsService;
+import io.devarium.infrastructure.auth.jwt.JwtConstants;
+import io.devarium.infrastructure.auth.jwt.JwtUtil;
+import io.devarium.infrastructure.auth.service.UserDetailsServiceImpl;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
@@ -15,6 +15,7 @@ import org.springframework.security.authentication.UsernamePasswordAuthenticatio
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.util.StringUtils;
 import org.springframework.web.filter.OncePerRequestFilter;
 
@@ -23,7 +24,7 @@ import org.springframework.web.filter.OncePerRequestFilter;
 public class JwtAuthenticationFilter extends OncePerRequestFilter {
 
     private final JwtUtil jwtUtil;
-    private final CustomUserDetailsService userDetailsService;
+    private final UserDetailsService userDetailsService;
 
     @Override
     protected void doFilterInternal(
@@ -36,7 +37,8 @@ public class JwtAuthenticationFilter extends OncePerRequestFilter {
         try {
             if (token != null && jwtUtil.isTokenValid(token)) {
                 String email = jwtUtil.extractEmail(token);
-                UserDetails userDetails = userDetailsService.loadUserByEmail(email);
+                UserDetails userDetails =
+                    ((UserDetailsServiceImpl) userDetailsService).loadUserByEmail(email);
 
                 Authentication authentication =
                     new UsernamePasswordAuthenticationToken(
