+++ infrastructure/src/main/java/io/devarium/infrastructure/auth/oauth2/config/WebClientConfig.java
@@ -1,4 +1,4 @@
-package io.devarium.infrastructure.oauth2.config;
+package io.devarium.infrastructure.auth.oauth2.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
