+++ infrastructure/src/main/java/io/devarium/infrastructure/auth/service/UserDetailsServiceImpl.java
@@ -1,7 +1,8 @@
-package io.devarium.infrastructure.security.userdetails;
+package io.devarium.infrastructure.auth.service;
 
 import io.devarium.core.domain.user.exception.UserErrorCode;
 import io.devarium.core.domain.user.exception.UserException;
+import io.devarium.infrastructure.auth.CustomUserDetails;
 import io.devarium.infrastructure.persistence.repository.UserJpaRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetails;
@@ -11,7 +12,7 @@ import org.springframework.stereotype.Service;
 
 @RequiredArgsConstructor
 @Service
-public class CustomUserDetailsService implements UserDetailsService {
+public class UserDetailsServiceImpl implements UserDetailsService {
 
     private final UserJpaRepository userJpaRepository;
